pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["kwg"]
description = "AI-powered web stack for digital humanities research"
license = "LICENSE"
name = "orca"
readme = "README"
version = "0.1.0"

[tool.poetry.dependencies]
aioboto3 = ">=13"
aiosqlite = ">=0.20"
cryptography = ">=43"
fastapi = ">=0.112"
gunicorn = ">=23"
icloudpy = ">=0.6,<0.7"
matplotlib = ">=3.9"
natsort = ">=8"
nltk = ">=3.9"
numpy = ">=2.1"
pandas = ">=2.2"
pillow = ">=10"
python = ">=3.12,<4"
python-dateutil = ">=2"
python-docx = ">=1.1,<1.2"
python-dotenv = ">=1"
python-slugify = {extras = ["unidecode"], version = ">=8"}
regex = ">=2024"
seaborn = ">=0.13"
scikit-learn = ">=1.5"
SQLAlchemy = {extras = ["asyncio"], version = ">=2,<3"}
Unidecode = ">=1.3"
uvicorn = ">=0.30"
Whoosh = ">=2.7"
tenacity = "^9.0.0"

[tool.poetry.group.dev.dependencies]
bandit = ">=1.7"
black = ">=24"
flake8 = ">=7"
flake8-bugbear = ">=24"
mypy = ">=1.11"
pytest = ">=8"
pytest-asyncio = ">=0.24"
pytest-cov = ">=5"
pytest-mock = ">=3"
sphinx = ">=7,<8"
toml-sort = ">=0.23"
tox = ">=4"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"  # silence deprec. warning
